Исправил ошибки из второго домашнего задания. (Убрал переобъявление серверных сессий, создал отдельный файл для взаимодействия с сервером)
