+++ VK-App/UI-Home Application/MainClassStructures.swift
