+++ VK-App/UI-Home Application/UpdatedPersonalPageViewController.swift
@@ -35,37 +35,10 @@ class PersonalPageViewController: UIViewController {
         "New Online course opens at GeekBrains"
     ]
     
-    
-    private class NetworkSession {
-        static let custom: Alamofire.Session = {
-            let configuration = URLSessionConfiguration.default
-            let sessionManager = Alamofire.Session(configuration: configuration)
-            return sessionManager
-        }()
-    }
-    
-    private func downloadUserData(){
-        let parameters: Parameters = [
-            "fields": "status,city,photo_50",
-            "access_token": Session.instance.app_token!,
-            "v": "5.101"
-        ]
-        
-        NetworkSession.custom.request("https://api.vk.com/method/users.get", parameters: parameters).responseJSON {
-            response in
-            
-            print(response.value)
-        }
-    }
-
-    
-    
-    
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        downloadUserData()
+        ServerTusks.instance.downloadUserData()
         
         headPageView.image = personData.photo
         headPageView.name = personData.last_name
